+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/timeline/element/message/TextRoomMessageTimelineElementView.kt
@@ -1,6 +1,8 @@
 package de.connect2x.messenger.compose.view.room.timeline.element.message
 
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ClipEntry
 import de.connect2x.messenger.compose.view.room.timeline.element.TimelineElementView
 import de.connect2x.messenger.compose.view.util.toClipEntry
@@ -18,6 +20,8 @@ class TextRoomMessageTimelineElementViewImpl : TextRoomMessageTimelineElementVie
     override suspend fun waitFor(element: Text) {
         // NO-OP (has default size)
     }
+
+    override fun isFocusable(): Boolean = true
 
     @Composable
     override fun createInTimeline(
@@ -39,16 +43,20 @@ class TextRoomMessageTimelineElementViewImpl : TextRoomMessageTimelineElementVie
     override fun createReplyInTimeline(
         holder: TimelineElementHolderViewModel,
         element: Text,
+        modifier: Modifier,
+        interactionSource: MutableInteractionSource,
     ) {
-        TextReplyInTimeline(holder, element)
+        TextReplyInTimeline(holder, element, modifier, interactionSource)
     }
 
     @Composable
     override fun createReplyInSendMessage(
         holder: TimelineElementHolderViewModel,
         element: Text,
+        modifier: Modifier,
+        interactionSource: MutableInteractionSource,
     ) {
-        TextReplyInSendMessage(holder, element)
+        TextReplyInSendMessage(holder, element, modifier, interactionSource)
     }
 
     @Composable