+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/room/Invite.kt
@@ -12,11 +12,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextOverflow
 import de.connect2x.messenger.compose.view.DI
-import de.connect2x.messenger.compose.view.common.Tooltip
-import de.connect2x.messenger.compose.view.common.TooltipText
-import de.connect2x.messenger.compose.view.get
+import de.connect2x.messenger.compose.view.common.Tooltip import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.theme.components
 import de.connect2x.messenger.compose.view.theme.components.ModalDialogContent
@@ -27,6 +26,7 @@ import de.connect2x.messenger.compose.view.theme.components.ThemedIconButton
 import de.connect2x.messenger.compose.view.theme.components.ThemedModalDialog
 import de.connect2x.messenger.compose.view.theme.components.ThemedProgressIndicator
 import de.connect2x.messenger.compose.view.theme.components.ThemedSelectableText
+import de.connect2x.messenger.compose.view.util.rovingFocusChild
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListElementViewModel
 
 interface InviteRoomListElement {
@@ -64,6 +64,7 @@ class InviteRoomListElementImpl : InviteRoomListElement {
                 ThemedIconButton(
                     style = MaterialTheme.components.commonIconButton,
                     onClick = { roomListElementViewModel.acceptInvitation() },
+                    modifier = Modifier.rovingFocusChild(),
                 ) {
                     Icon(Icons.Default.Check, i18n.invitationAccept())
                 }
@@ -74,6 +75,7 @@ class InviteRoomListElementImpl : InviteRoomListElement {
                 ThemedIconButton(
                     style = MaterialTheme.components.commonIconButton,
                     onClick = { showReject = true },
+                    modifier = Modifier.rovingFocusChild(),
                 ) {
                     Icon(Icons.Default.Close, i18n.invitationReject())
                 }
@@ -123,10 +125,8 @@ class InviteRoomListElementImpl : InviteRoomListElement {
     fun RoomInviterUserInfo(inviterNameOrUserId: String) {
         val needsTooltip = remember { mutableStateOf(false) }
         Tooltip(
-            tooltip = {
-                TooltipText(inviterNameOrUserId)
-            },
-            enabled = needsTooltip.value
+            enabled = needsTooltip.value,
+            tooltip = { Text(inviterNameOrUserId) }
         ) {
             Text(
                 inviterNameOrUserId,